+++ src/components/PageTitleTwo.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="w-full bg-white flex flex-col fixed z-40 border-b-1 border-gray-primary border-opacity-15 pb-12">
+  <div v-bind:class="customClass" class="w-full flex flex-col fixed z-40">
 
     <!-- HEADER -->
     <div class="flex flex-row w-full justify-between items-center pt-16 px-5">
@@ -12,7 +12,7 @@
         <slot name="title"/>
       </div>
 
-      <div class="w-2/12">
+      <div class="w-2/12 flex flex-row justify-end">
         <slot name="rightAction"/>
       </div>
     </div>
@@ -22,7 +22,36 @@
 
 <script>
 export default {
-name: "PageTitleTwo"
+  name: "PageTitleTwo",
+  props: {
+    backgroundColor:{
+      type: String,
+      default: 'bg-white'
+    },
+    bottomBorder: {
+      type: Boolean,
+      default: true
+    },
+    bottomPadding: {
+      type: Number,
+      default: 12
+    }
+  },
+  computed: {
+    customClass: function (){
+
+      let value = '';
+
+      value += this.backgroundColor;
+      value += ' pb-' + this.bottomPadding;
+
+      if (this.bottomBorder) {
+        value += ' border-b-1 border-gray-primary border-opacity-15'
+      }
+
+      return value;
+    }
+  }
 }
 </script>