+++ src/router/teacher_assignment_routes.js
@@ -2,9 +2,11 @@ import AssignmentShow from "@/views/teachers/TeacherAssignments/Show/Index";
 import AssignmentDetails from "@/views/teachers/TeacherAssignments/Show/Components/AssignmentDetails";
 import AlternateDashboardLayout from "@/views/layout/AlternateDashboardLayout";
 import AssignmentCreate from "@/views/teachers/TeacherAssignments/Create/Index";
+import AssignmentMark from "@/views/teachers/TeacherAssignments/Mark/Index";
 import AssignmentForm from "@/views/teachers/TeacherAssignments/Create/Components/AssignmentForm";
-import QuestionForm from "@/views/teachers/TeacherAssignments/Create/Components/QuestionForm";
-import QuestionManualDesc from "@/views/teachers/TeacherAssignments/Create/Components/QuestionManualDesc";
+import AssignmentMarkDetails from '@/views/teachers/TeacherAssignments/Mark/Components/AssignmentDetails';
+import AssignmentMarkFeedback from '@/views/teachers/TeacherAssignments/Mark/Components/AssignmentFeedback';
+
 
 const teacherAccessControlMeta = {
     checkAuth: 'true',
@@ -65,6 +67,27 @@ export default {
                     component: AssignmentDetails,
                     meta: teacherAccessControlMeta,
                     props: true,
+                },
+                {
+                    path: 'marking',
+                    component: AssignmentMark,
+                    children: [
+                        {
+                            path: '',
+                            name: 'teacher.assignments.marking.details',
+                            component: AssignmentMarkDetails,
+                            meta: teacherAccessControlMeta,
+                            props: true,
+                        },
+                        {
+                            path: 'feedback',
+                            name: 'teacher.assignments.marking.feedback',
+                            component: AssignmentMarkFeedback,
+                            meta: teacherAccessControlMeta,
+                            props: true,
+                        }
+
+                    ]
                 }
             ]
         },