+++ src/views/teachers/TeacherAssignments/Mark/Components/AssignmentAnswerPreview.vue
@@ -1,0 +1,13 @@
+<template>
+
+</template>
+
+<script>
+export default {
+name: "AssignmentAnswerPreview"
+}
+</script>
+
+<style scoped>
+
+</style>